Replace 22 incremental migrations with single initial schema migration
Consolidates all migrations (0000-0021) into one 0001_initial_schema.py
that creates the complete database schema in a single pass. Fixes fresh
install failures where incremental migrations assumed tables existed from
Base.metadata.create_all.

Also switches alembic env.py from asyncpg to psycopg2 for migrations,
avoiding strict type handling issues with DDL operations.

Existing installs: entrypoint auto-stamps at head when tables exist
without alembic tracking.
