--- backend/alembic/versions/0009_add_user_and_refresh_token_tables.py
@@ -1,49 +1,0 @@
-"""Add user and refresh_token tables."""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-
-revision = "0009"
-down_revision = "0008"
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    user_role = sa.Enum("admin", "viewer", name="user_role")
-    user_role.create(op.get_bind(), checkfirst=True)
-
-    op.create_table(
-        "users",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column("username", sa.String(150), unique=True, nullable=False),
-        sa.Column("password_hash", sa.String(255), nullable=False),
-        sa.Column("role", user_role, nullable=False),
-        sa.Column("is_active", sa.Boolean, nullable=False, server_default=sa.text("true")),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False, server_default=sa.func.now()),
-        sa.Column("updated_at", sa.DateTime(timezone=True), nullable=False, server_default=sa.func.now()),
-        if_not_exists=True,
-    )
-
-    op.create_table(
-        "refresh_tokens",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column("user_id", UUID(as_uuid=True), sa.ForeignKey("users.id", ondelete="CASCADE"), nullable=False),
-        sa.Column("token_hash", sa.String(64), unique=True, nullable=False),
-        sa.Column("family_id", UUID(as_uuid=True), nullable=False),
-        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("revoked", sa.Boolean, nullable=False, server_default=sa.text("false")),
-        sa.Column("created_at", sa.DateTime(timezone=True), nullable=False, server_default=sa.func.now()),
-        if_not_exists=True,
-    )
-
-    op.create_index("ix_refresh_tokens_user_id", "refresh_tokens", ["user_id"])
-    op.create_index("ix_refresh_tokens_family_id", "refresh_tokens", ["family_id"])
-
-
-def downgrade() -> None:
-    op.drop_index("ix_refresh_tokens_family_id", table_name="refresh_tokens")
-    op.drop_index("ix_refresh_tokens_user_id", table_name="refresh_tokens")
-    op.drop_table("refresh_tokens")
-    op.drop_table("users")
-    sa.Enum(name="user_role").drop(op.get_bind(), checkfirst=True)
